Restore private API surface for GPUI panic helpers
The implementation had exposed `GpuiHarness::augmented_panic_message`
and `GpuiHarness::write_stderr_diagnostic_to` as `pub` solely to let the
integration tests call them directly. Both contradict the ExecPlan's
"no new public types, traits, or functions" constraint, are unused by
any downstream caller, and are not part of the user-facing harness
contract. Drop `pub` from both.

Tests that previously poked the private surface are reshaped:

- `augmented_message_includes_scenario_name_for_{string,str,opaque_any}_payload`
  now route through `catch_scenario_panic`, triggering each payload
  variant with `panic_any` or a bare `panic!(literal)` so the harness's
  panic flow is exercised end-to-end. The existing snapshots already
  matched the augmented message format and remain unchanged.
- `stderr_write_failure_is_non_fatal` moves into the crate-internal
  `#[cfg(test)] mod tests` (renamed
  `write_stderr_diagnostic_to_returns_err_on_io_failure`) so it can
  call the now-private writer without re-exposing it. The shared
  `BrokenPipeWriter` sink lives next to the test.

Sync the ExecPlan's "Interfaces and dependencies" section to describe
the actual helpers (`augmented_panic_message`, `record_panic_event`,
`write_stderr_diagnostic_to`, `ContextCleanup`) and to document that
the lone unit test for the I/O primitive lives in the internal
`mod tests` rather than the integration test crate.
